Move util::str::parse_legacy_font_size to script::dom::htmlfontelement fixes #9638
